refactor(components/): Move `switchChain` to `walletUtils.ts`
